dbe/coreAPI/requirements.txt
asgiref==3.5.2
Django==3.2.14
django-cors-headers==3.5.0
django-dotenv==1.4.2
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
